mise a jour

+++ src/main/java/com/afh/gescomp/model/secondary/Role.java
@@ -1,6 +1,7 @@
+/*
 package com.afh.gescomp.model.secondary;
 
-import jakarta.persistence.*;
+import javax.persistence.*;
 import lombok.*;
 
 @Getter
@@ -12,9 +13,10 @@ import lombok.*;
 public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Enumerated(EnumType.STRING)
     @Column(length = 20)
     private ERole name;
 }
+*/
