+++ src/main/java/com/afh/gescomp/repository/secondary/RoleRepository.java
@@ -1,3 +1,4 @@
+/*
 package com.afh.gescomp.repository.secondary;
 
 import com.afh.gescomp.model.secondary.ERole;
@@ -6,7 +7,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RoleRepository  extends JpaRepository<Role, Integer> {
+public interface RoleRepository  extends JpaRepository<Role, Long> {
 
     Optional<Role> findByName(ERole name);
 }
+*/
